There's been lots of code proliferation and its starting to get hard to extend. So need to organize / refactor a bit.
